requirements.txt
requests>=2.7.0
ConfigMaster>=2.3.4
Flask>=0.10.1
itsdangerous>=0.24
Jinja2>=2.8
MarkupSafe>=0.23
psutil>=3.1.1
PyYAML>=3.11
Werkzeug>=0.10.4
redis>=2.10.3
git+https://github.com/benoitc/gunicorn.git
Flask-Cache>=0.13.1
Flask-Migrate>=1.5.0
Flask-Script>=2.0.5
Flask-SQLAlchemy>=2.0
Flask-SQLAlchemy-Cache>=0.1.2
psycopg2
git+https://github.com/SunDwarf/python-pgpdump.git@suport-ecc # Not a typo